tools/willy_re/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "willy-re"
version = "0.1.0"
description = "Reverse engineering toolkit for Willy Werkel / Mulle Meck Director games"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1",
    "Pillow>=10.0",
    "tqdm>=4.65",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "ruff>=0.1"]

[project.scripts]
willy-re = "willy_re.cli:main"

[tool.setuptools.packages.find]
where = ["."]

[tool.ruff]
line-length = 100
target-version = "py310"
